// src/domain/constants/codes.enum.ts
export enum CodesEnum {
    BUY_CORN_SUCCESS_BUY = "BC-SUC-200-001",
    BUY_CORN_NOT_FOUND_CUSTOMER = "BC-ERR-404-001",
    BUY_CORN_SERVER_ERROR = "BC-ERR-500-001",
    BUY_CORN_RATE_LIMIT_EXCEEDED = "BC-ERR-429-001",
    CREATE_CUSTOMER_SUCCESS = "CC-SUC-201-001",
    CREATE_CUSTOMER_ALREADY_EXISTS = "CC-ERR-409-001",
    CREATE_CUSTOMER_SERVER_ERROR = "CC-ERR-500-001",
    LOGIN_SUCCESS = "LG-SUC-200-001",
    LOGIN_INVALID_CREDENTIALS = "LG-ERR-401-001",
    LOGIN_SERVER_ERROR = "LG-ERR-500-001",
}